third_party/blink/web_tests/platform/linux/fast/forms/datetimelocal-multiple-fields/datetimelocal-multiple-fields-preserve-value-after-history-back-expected.txt

Checks input value is preserved after backward and forward

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Went back to a page. Checking a form control which had a full value:
PASS "12012-10-09T01:23" is "12012-10-09T01:23"
Went back to a page. Checking a form control which had a partial value:
PASS "01/dd/1999, --:00 --" is "01/dd/1999, --:00 --"
Went forward to a page. Checking a form control which had a full value:
PASS "0001-11-17T23:55" is "0001-11-17T23:55"
Went forward to a page. Checking a form control which had a partial value:
PASS "mm/31/yyyy, 12:-- PM" is "mm/31/yyyy, 12:-- PM"


PASS successfullyParsed is true

TEST COMPLETE
